Refactor Matrix lookup logic into separate file
Moved the lookupExisting method from Matrix.swift to a new Matrix+Lookup.swift extension file for better separation of concerns. Also removed related methods from SyncGroup that depended on the relocated lookup logic.
